LoRa stack use float instead of float for codesize optimization

+++ kernel/protocols/lorawan/linkwan/region/RegionCN470A.c
@@ -36,6 +36,7 @@ Jaeckle ( STACKFORCE )
 #include "debug.h"
 
 #include "linkwan.h"
+#include "precision.h"
 
 // Definitions
 #define CHANNELS_MASK_SIZE 1
@@ -485,7 +486,7 @@ void RegionCN470AComputeRxWindowParameters(int8_t            datarate,
                                            uint32_t          rxError,
                                            RxConfigParams_t *rxConfigParams)
 {
-    double tSymbol = 0.0;
+    DECIMAL tSymbol = 0.0f;
 
     rxConfigParams->Datarate  = datarate;
     rxConfigParams->Bandwidth = GetBandwidth(datarate);
@@ -551,9 +552,8 @@ bool RegionCN470ARxConfig(RxConfigParams_t *rxConfig, int8_t *datarate)
     // Radio configuration
     if (dr == DR_7) {
         modem = MODEM_FSK;
-        Radio.SetRxConfig(modem, 50e3, phyDr * 1e3, 0, 83.333e3, 5,
-                          rxConfig->WindowTimeout, false, 0, true, 0, 0, false,
-                          rxConfig->RxContinuous);
+        Radio.SetRxConfig( modem, 50e3f, phyDr * 1e3f, 0, 83.333e3f, 5, rxConfig->WindowTimeout, false, 0, true, 0, 0, false,
+                           rxConfig->RxContinuous );
     } else {
         modem = MODEM_LORA;
         Radio.SetRxConfig(modem, rxConfig->Bandwidth, phyDr, 1, 0, 8,
@@ -611,12 +611,10 @@ bool RegionCN470ATxConfig(TxConfigParams_t *txConfig, int8_t *txPower,
     if (txConfig->Datarate == DR_7) {
         // High Speed FSK channel
         modem = MODEM_FSK;
-        Radio.SetTxConfig(modem, phyTxPower, 25e3, bandwidth, phyDr * 1e3, 0, 5,
-                          false, true, 0, 0, false, 3e3);
+        Radio.SetTxConfig( modem, phyTxPower, 25e3f, bandwidth, phyDr * 1e3f, 0, 5, false, true, 0, 0, false, 3e3f );
     } else {
         modem = MODEM_LORA;
-        Radio.SetTxConfig(modem, phyTxPower, 0, bandwidth, phyDr, 1,
-                          preambleLen, false, true, 0, 0, iqInverted, 3e3);
+        Radio.SetTxConfig( modem, phyTxPower, 0, bandwidth, phyDr, 1, preambleLen, false, true, 0, 0, iqInverted, 3e3f );
     }
     // Setup maximum payload lenght of the radio driver
     Radio.SetMaxPayloadLength(modem, txConfig->PktLen);
